+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import MiniAppWrapper from "@/components/MiniAppWrapper";
-import { miniAppConfig } from "@/lib/config";
 import "./globals.css";
 
 const geistSans = Geist({
@@ -17,22 +16,6 @@ const geistMono = Geist_Mono({
 export const metadata: Metadata = {
   title: "Farcaster Lore - GM Farcaster Casts",
   description: "Explore casts featured on GM Farcaster episodes through time",
-  other: {
-    "fc:frame": JSON.stringify({
-      "version": "next",
-      "imageUrl": miniAppConfig.imageUrl(),
-      "button": {
-        "title": "Travel Through Farcaster History",
-        "action": {
-          "type": "launch_miniapp",
-          "name": "Farcaster Lore",
-          "url": miniAppConfig.homeUrl(),
-          "splashImageUrl": miniAppConfig.splashImageUrl(),
-          "splashBackgroundColor": miniAppConfig.splashBackgroundColor
-        }
-      }
-    })
-  }
 };
 
 export default function RootLayout({
