+++ src/app/page.tsx
@@ -1,300 +1,33 @@
-'use client';
+import type { Metadata } from 'next';
+import { miniAppConfig } from '@/lib/config';
+import HomeClient from '@/components/HomeClient';
 
-import { useState, useEffect } from 'react';
-import { useMiniApp } from '@neynar/react';
-import { EnrichedCast, SearchFilters } from '@/lib/types';
-import { searchCasts, getAllCasts, validateAndCleanCasts } from '@/lib/api';
-import CastGrid from '@/components/CastGrid';
-import { useDevMode } from '@/lib/devMode';
+export async function generateMetadata(): Promise<Metadata> {
+  const miniappEmbed = {
+    "version": "next",
+    "imageUrl": miniAppConfig.imageUrl(),
+    "button": {
+      "title": "Travel Through Farcaster History",
+      "action": {
+        "type": "launch_miniapp",
+        "name": "Farcaster Lore",
+        "url": miniAppConfig.homeUrl(),
+        "splashImageUrl": miniAppConfig.splashImageUrl(),
+        "splashBackgroundColor": miniAppConfig.splashBackgroundColor
+      }
+    }
+  };
 
-import DatePicker from '@/components/DatePicker';
-import UserSearch from '@/components/UserSearch';
-import Link from 'next/link';
-
-const CASTS_PER_PAGE = 20;
+  return {
+    title: "Farcaster Lore - GM Farcaster Casts",
+    description: "Explore casts featured on GM Farcaster episodes through time",
+    other: {
+      "fc:miniapp": JSON.stringify(miniappEmbed),
+      "fc:frame": JSON.stringify(miniappEmbed)
+    }
+  };
+}
 
 export default function Home() {
-  const { context } = useMiniApp();
-  const { isDevMode, devContext } = useDevMode();
-  
-  // Use dev mode context if available, otherwise use real miniapp context
-  const effectiveContext = isDevMode ? devContext : { context };
-  
-  const [casts, setCasts] = useState<EnrichedCast[]>([]);
-  const [totalCasts, setTotalCasts] = useState(0);
-  const [isLoading, setIsLoading] = useState(true);
-  const [isLoadingMore, setIsLoadingMore] = useState(false);
-  const [activeFilter, setActiveFilter] = useState('all');
-  const [selectedDate, setSelectedDate] = useState<string | null>(null);
-  const [selectedUser, setSelectedUser] = useState<{ username?: string; fid?: number } | null>(null);
-  const [isDatePickerOpen, setIsDatePickerOpen] = useState(false);
-  const [isUserSearchOpen, setIsUserSearchOpen] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  // Load initial casts
-  useEffect(() => {
-    loadInitialCasts();
-  }, []);
-
-  const loadInitialCasts = async () => {
-    try {
-      setIsLoading(true);
-      setError(null);
-      
-      const response = await getAllCasts(CASTS_PER_PAGE, 0);
-      const cleanedCasts = validateAndCleanCasts(response.casts);
-      setCasts(cleanedCasts);
-      setTotalCasts(response.total);
-      setIsLoading(false);
-    } catch (error) {
-      console.error('Error loading casts:', error);
-      setError('Failed to load casts. Please try again.');
-      setIsLoading(false);
-    }
-  };
-
-  const loadMore = async () => {
-    if (isLoadingMore) return;
-    
-    try {
-      setIsLoadingMore(true);
-      
-      const filters: SearchFilters = {
-        limit: CASTS_PER_PAGE,
-        offset: casts.length,
-      };
-      
-      // Add current filters
-      if (selectedDate) filters.date = selectedDate;
-      if (selectedUser?.username) {
-        filters.author = selectedUser.username;
-      } else if (selectedUser?.fid) {
-        filters.fid = selectedUser.fid;
-      }
-      
-      const response = await searchCasts(filters);
-      const cleanedCasts = validateAndCleanCasts(response.casts);
-      
-      setCasts(prev => [...prev, ...cleanedCasts]);
-    } catch (error) {
-      console.error('Error loading more casts:', error);
-      setError('Failed to load more casts. Please try again.');
-    } finally {
-      setIsLoadingMore(false);
-    }
-  };
-
-  const handleFilterChange = async (filter: string) => {
-    setActiveFilter(filter);
-    
-    if (filter === 'all') {
-      // Clear all filters when going back to "All Casts"
-      setSelectedDate(null);
-      setSelectedUser(null);
-      await loadInitialCasts();
-    }
-  };
-
-  const handleDateSelect = async (date: string | null) => {
-    console.log('[DatePicker] Selected date:', date);
-    setSelectedDate(date);
-    setActiveFilter(date ? 'pick-date' : 'all');
-    
-    // Clear user search when switching to date filter
-    setSelectedUser(null);
-    
-    try {
-      setIsLoading(true);
-      setError(null);
-      
-      const filters: SearchFilters = {
-        limit: CASTS_PER_PAGE,
-        offset: 0,
-      };
-      
-      if (date) {
-        filters.date = date;
-      }
-      console.log('[DatePicker] Filters for search:', filters);
-      
-      const response = await searchCasts(filters);
-      console.log('[DatePicker] API response:', response);
-      const cleanedCasts = validateAndCleanCasts(response.casts);
-      setCasts(cleanedCasts);
-      setTotalCasts(response.total);
-    } catch (error) {
-      console.error('Error filtering by date:', error);
-      setError('Failed to filter by date. Please try again.');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleDatePickerOpen = () => {
-    setIsDatePickerOpen(true);
-  };
-
-  const handleUserSearchOpen = () => {
-    setIsUserSearchOpen(true);
-  };
-
-  const handleUserSelect = async (username: string | null, fid: number | null) => {
-    setSelectedUser(username ? { username } : fid ? { fid } : null);
-    setActiveFilter((username || fid) ? 'pick-user' : 'all');
-    
-    // Clear date filter when switching to user search
-    setSelectedDate(null);
-    
-    try {
-      setIsLoading(true);
-      setError(null);
-      
-      const filters: SearchFilters = {
-        limit: CASTS_PER_PAGE,
-        offset: 0,
-      };
-      
-      if (username) {
-        filters.author = username;
-      } else if (fid) {
-        filters.fid = fid;
-      }
-      
-      const response = await searchCasts(filters);
-      const cleanedCasts = validateAndCleanCasts(response.casts);
-      setCasts(cleanedCasts);
-      setTotalCasts(response.total);
-    } catch (error) {
-      console.error('Error filtering by user:', error);
-      setError('Failed to filter by user. Please try again.');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const hasMore = casts.length < totalCasts;
-
-  return (
-    <main className="min-h-screen bg-gray-50">
-      <header className="bg-white shadow-sm border-b border-gray-200">
-        <div className="max-w-7xl mx-auto px-4 py-6">
-          {/* Row 1: Title & Subtitle (full width) */}
-          <div className="text-center mb-6">
-            <h1 className="text-2xl font-bold text-gray-900">
-              Farcaster Lore
-            </h1>
-            <p className="text-gray-600 mt-1">
-              Travel through Farcaster history with every cast mentioned on{' '}
-              <a 
-                href="https://www.youtube.com/@gmfarcaster"
-                target="_blank" 
-                rel="noopener noreferrer"
-                className="text-purple-600 hover:text-purple-800 font-medium"
-              >
-                GM Farcaster
-              </a>
-            </p>
-            {isDevMode && (
-              <div className="mt-2 inline-block bg-yellow-100 text-yellow-800 px-3 py-1 rounded-full text-sm font-medium">
-                🧪 Dev Mode - Test FID: {effectiveContext.context?.user?.fid}
-              </div>
-            )}
-          </div>
-
-          {/* Row 2: Main Navigation (prominent) */}
-          <div className="flex gap-3 justify-center mb-4">
-            {effectiveContext.context?.user?.fid && (
-              <Link 
-                href={isDevMode ? "/my-references?dev=true" : "/my-references"}
-                className="px-6 py-3 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors font-medium"
-              >
-                📍 My References
-              </Link>
-            )}
-            <Link 
-              href="/leaderboard"
-              className="px-6 py-3 bg-purple-700 text-white rounded-lg hover:bg-purple-800 transition-colors font-medium"
-            >
-              🏆 Leaderboard
-            </Link>
-          </div>
-
-          {/* Row 3: Filter Navigation (less prominent) */}
-          <div className="flex gap-2 justify-center">
-            <button 
-              onClick={() => handleFilterChange('all')}
-              className={`px-4 py-2 rounded-lg transition-colors text-sm ${
-                activeFilter === 'all' 
-                  ? 'bg-purple-100 text-purple-800 border border-purple-300' 
-                  : 'bg-gray-100 text-gray-600 hover:bg-gray-200'
-              }`}
-            >
-              All Casts
-            </button>
-            <button 
-              onClick={handleUserSearchOpen}
-              className={`px-4 py-2 rounded-lg transition-colors text-sm ${
-                activeFilter === 'pick-user' 
-                  ? 'bg-purple-100 text-purple-800 border border-purple-300' 
-                  : 'bg-gray-100 text-gray-600 hover:bg-gray-200'
-              }`}
-            >
-              Pick a User
-            </button>
-            <button 
-              onClick={handleDatePickerOpen}
-              className={`px-4 py-2 rounded-lg transition-colors text-sm ${
-                activeFilter === 'pick-date' 
-                  ? 'bg-purple-100 text-purple-800 border border-purple-300' 
-                  : 'bg-gray-100 text-gray-600 hover:bg-gray-200'
-              }`}
-            >
-              Pick a Date
-            </button>
-          </div>
-        </div>
-      </header>
-
-
-
-      <DatePicker
-        isOpen={isDatePickerOpen}
-        onClose={() => setIsDatePickerOpen(false)}
-        onDateSelect={handleDateSelect}
-      />
-
-      <UserSearch
-        isOpen={isUserSearchOpen}
-        onClose={() => setIsUserSearchOpen(false)}
-        onUserSelect={handleUserSelect}
-      />
-
-      {error && (
-        <div className="max-w-7xl mx-auto px-4 py-4">
-          <div className="bg-red-50 border border-red-200 rounded-lg p-4">
-            <p className="text-red-800">{error}</p>
-            <button 
-              onClick={loadInitialCasts}
-              className="mt-2 text-red-600 hover:text-red-800 underline"
-            >
-              Try again
-            </button>
-          </div>
-        </div>
-      )}
-
-      {isLoading ? (
-        <div className="flex items-center justify-center py-12">
-          <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-purple-600"></div>
-        </div>
-      ) : (
-        <CastGrid
-          casts={casts}
-          onLoadMore={loadMore}
-          hasMore={hasMore}
-          isLoading={isLoadingMore}
-        />
-      )}
-    </main>
-  );
+  return <HomeClient />;
 }
